moved synapse update code generation into module
